[update] Stageの調整。
