--- Assets/Scripts/Stage/Maruoka.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8a6e8ee44579b494bae815fce9d371e2
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
