--- Assets/Scripts/Stage/Maruoka/Converter.cs
@@ -1,13 +1,0 @@
-// 日本語対応
-using UnityEngine;
-
-namespace TeamB_TD
-{
-    namespace Stage
-    {
-        public static class Converter
-        {
-
-        }
-    }
-}
