--- Assets/Scripts/Stage/Maruoka/Converter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9d1e4e837ee09d949bcebfd16d7a8007
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
