--- Assets/Scripts/Stage/ProtoCell.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 559c894376b908f4c8969f2f1f434900
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
