--- Assets/Scripts/Stage/ProtoStageGenerator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ec08dd31e321fe842ada094a93e3c2e3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
