+++ Assets/Scripts/Stage/StageController.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 namespace TeamB_TD
 {
-    namespace Stage
+    namespace StageManagement
     {
         public class StageController : MonoBehaviour
         {
